add existing save file case

+++ src/rw.py
@@ -1,10 +1,16 @@
 import json
+import os
 
 RESOURCE_PATH = f'.\\resources\\'
 
 def save_file(name, extension='.json', data=''):
-    with open(RESOURCE_PATH + name + extension, mode='w') as file:
-        file.write(data)
+    if os.path.isfile(RESOURCE_PATH + name + extension):
+        os.remove(RESOURCE_PATH + name + extension)
+        with open(RESOURCE_PATH + name + extension, mode='w') as file:
+            file.write(data)
+    else:
+        with open(RESOURCE_PATH + name + extension, mode='w+') as file:
+            file.write(data)
 
 def read_file(name, extension='.json'):
     with open(RESOURCE_PATH + name + extension) as file_data:
